refactor(dispatch): extract envelope_prepare.py leaf module (PR-2 monolith split)
Moves _prepare, _record_integrity, _verify_role_application out of
dispatch_envelope.py unchanged. dispatch_envelope.py re-exports and
calls all three at bare name so existing patch-string couplings keep
binding — the callers (_prepare/_record_integrity/_verify_role_application
call sites, including _govern's own call) stay in the facade.

Census fixture (tests/data/dispatch_envelope_census.json) regenerated:
family gains envelope_prepare, the three symbols relabel from
dispatch_envelope:: to envelope_prepare:: in layer4_functions. Coupling
count is unchanged (0 exposed patch-sites on these three symbols).

Dispatch-ID: 20260803-221903-pr2-envelope-prepare
